tools/client/requirements.txt
websocket-client
argparse
